+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Upgrade to class validator
 - [Issue #202] Switch to .env from local.config.json
 - [Issue #218] Remove and replaced loggging with ts-tiny-log
+- Hardened types to replace many `any` members and parameters
 
 ### Additions